+++ src/TCN75.h
@@ -4,16 +4,16 @@
  
 class TCN75 {
     public:
-        TCN75(I2C* i2c, uint16_t address);
+        TCN75(I2C* i2c, uint8_t address);
 
         void sleep();
         void wake();
 
-        uint16_t temperature();
+        int16_t temperature();
     
     private:
         I2C* i2c;
-        uint16_t address;
+        uint8_t address;
     
     private:
         const char ConfigPointer = 0x01;
